refactor(collector): extract storage.go and config_loader.go
Move StoreMetricsWithCopy into storage.go and config resolution
functions (LoadProbeConfigs, EnsureProbeConfig, etc.) into
config_loader.go. base.go now contains only interface definitions.

Part of #75.
